Moved Favorites button to a separate component file and linked it to favorites.js and view-space.js
