+++ src/components/index.js
@@ -6,5 +6,6 @@ import SurveyForm from './survey-form';
 import Profile from './profile';
 import Topbar from './topbar';
 import AddForm from './add-form'
+import FavoritesButton from './favoritesButton';
 
-export { Ambience, Homebar, Popularity, StudySpaceCard, SurveyForm, Profile, Topbar, AddForm };
+export { Ambience, FavoritesButton, Homebar, Popularity, StudySpaceCard, SurveyForm, Profile, Topbar, AddForm };
